Moves accel-mma84 to a library.
