+++ accel-mma84/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "accel-mma84"
-version = "0.1.0"
+version = "0.1.0-beta.1"
 authors = ["Tim Ryan <tim@tessel.io>"]
 description = "Driver for an MMA8452Q 3-axis accelerometer."
 license = "MIT"
